requirements.txt
aniso8601==8.1.1
asgiref==3.3.4
certifi==2020.12.5
chardet==3.0.4
Django==2.2.23
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-jwt==0.3.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-mptt-admin==2.1.0
django-richtextfield==1.6.1
django-sort-order-field==1.2
graphene==3.0b7
graphene-django==3.0.0b7
graphql-core==3.1.4
graphql-relay==3.1.0
idna==2.10
ippanel==1.1.0
Pillow==8.2.0
promise==2.3
PyJWT==2.0.1
pytz==2021.1
requests==2.23.0
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.25.11
